Split settings into response delivery and form options
