+++ components/form-builder/app/Settings.tsx
@@ -7,14 +7,12 @@ import Markdown from "markdown-to-jsx";
 import { useTemplateStore, clearTemplateStore } from "../store";
 import {
   Button,
-  Input,
   ConfirmFormDeleteDialog,
   useDialogRef,
   Dialog,
   DownloadFileButton,
 } from "./shared";
 import { useDeleteForm } from "../hooks";
-import { isValidGovEmail } from "@lib/validation";
 
 const FormDeleted = () => {
   const { t } = useTranslation("form-builder");
@@ -75,30 +73,8 @@ const HintText = ({ id, children }: { id: string; children?: JSX.Element | strin
   );
 };
 
-const InvalidEmailError = ({ id, isActive }: { id: string; isActive: boolean }) => {
-  const { t, i18n } = useTranslation("form-builder");
-
-  return (
-    <div id={id} className="mt-2 mb-2" role="alert">
-      {isActive && (
-        <>
-          <h2 className="text-red text-sm font-bold pb-1">{t("settingsInvalidEmailAlertTitle")}</h2>
-          <div className="bg-red-100 w-3/5 text-sm p-2">
-            <span>{t("settingsInvalidEmailAlertDesc1")}</span>
-            <br />
-            <a href={`/${i18n.language}/form-builder/support`} target="_blank" rel="noreferrer">
-              {t("contactSupport")}
-            </a>
-            <span> {t("settingsInvalidEmailAlertDesc2")}</span>
-          </div>
-        </>
-      )}
-    </div>
-  );
-};
-
 export const Settings = () => {
-  const { t, i18n } = useTranslation("form-builder");
+  const { t } = useTranslation("form-builder");
   const { handleDelete } = useDeleteForm();
   const [formDeleted, setFormDeleted] = useState(false);
   const [error, setError] = useState(false);
@@ -107,93 +83,16 @@ export const Settings = () => {
   const { deleteconfirm, downloadconfirm } = router.query;
   const [showConfirm, setShowConfirm] = useState(deleteconfirm || false);
 
-  const { id, initialize, email, updateField, isPublished } = useTemplateStore((s) => ({
+  const { id, initialize, isPublished } = useTemplateStore((s) => ({
     id: s.id,
     initialize: s.initialize,
-    email: s.deliveryOption?.emailAddress,
-    updateField: s.updateField,
     isPublished: s.isPublished,
   }));
   const { status } = useSession();
-  const [inputEmail, setInputEmail] = useState(email ?? "");
-  const [IsInvalidEmailErrorActive, setIsInvalidEmailErrorActive] = useState(false);
-
-  const handleEmailChange = (email: string) => {
-    setInputEmail(email);
-
-    const completeEmailAddressRegex =
-      /^([a-zA-Z0-9!#$%&'*+-/=?^_`{|}~.])+@([a-zA-Z0-9-.]+)\.([a-zA-Z0-9]{2,})+$/;
-
-    // We want to make sure the email address is complete before validating it
-    if (!completeEmailAddressRegex.test(email)) {
-      setIsInvalidEmailErrorActive(false);
-      return;
-    }
-
-    if (isValidGovEmail(email)) {
-      setIsInvalidEmailErrorActive(false);
-      updateField(`deliveryOption.emailAddress`, email);
-    } else {
-      setIsInvalidEmailErrorActive(true);
-    }
-  };
 
   return (
     <>
       <h1 className="visually-hidden">{t("formSettings")}</h1>
-
-      {isPublished && (
-        <div className="mb-10">
-          <Label htmlFor="response-delivery">{t("settingsResponseTitle")}</Label>
-          <HintText id="response-delivery-hint-1">{t("settingsResponseHint1")}</HintText>
-          <HintText id="response-delivery-hint-2">{t("settingsResponseHint2")}</HintText>
-          <div className="mt-4 mb-4 p-4 bg-purple-200 text-sm inline-block">
-            {t("settingsResponseNotePublished")}
-            <a
-              href={`/${i18n.language}/form-builder/support`}
-              className="ml-2"
-              target="_blank"
-              rel="noreferrer"
-            >
-              {t("contactSupport")}
-            </a>
-            .
-          </div>
-          <div>{inputEmail}</div>
-        </div>
-      )}
-
-      {!isPublished && (
-        <div className="mb-10">
-          <Label htmlFor="response-delivery">{t("settingsResponseTitle")}</Label>
-          <HintText id="response-delivery-hint-1">{t("settingsResponseHint1")}</HintText>
-          <HintText id="response-delivery-hint-2">{t("settingsResponseHint2")}</HintText>
-          <div className="mt-4 p-4 bg-purple-200 text-sm inline-block">
-            {t("settingsResponseNote")}
-            <a
-              href={`/${i18n.language}/form-builder/support`}
-              className="ml-2"
-              target="_blank"
-              rel="noreferrer"
-            >
-              {t("contactSupport")}
-            </a>
-            .
-          </div>
-          <InvalidEmailError id="invalidEmailError" isActive={IsInvalidEmailErrorActive} />
-
-          <div className="block font-bold mb-1 text-sm">{t("settingsResponseEmailTitle")}</div>
-          <Input
-            id="response-delivery"
-            isInvalid={IsInvalidEmailErrorActive}
-            describedBy="response-delivery-hint-1 response-delivery-hint-2 invalidEmailError"
-            value={inputEmail}
-            theme={IsInvalidEmailErrorActive ? "error" : "default"}
-            className="w-3/5"
-            onChange={(e) => handleEmailChange(e.target.value)}
-          />
-        </div>
-      )}
 
       <div id="download-form" className="mb-6">
         <Label htmlFor="download">{t("formDownload.title")}</Label>
