+++ components/form-builder/app/index.ts
@@ -1,3 +1,4 @@
+export { ResponseDelivery } from "./ResponseDelivery";
 export { Settings } from "./Settings";
 export { LeftNavigation } from "./navigation/LeftNavigation";
 export { PreviewNavigation } from "./navigation/PreviewNavigation";
