+++ components/form-builder/hooks/useElementOptions.tsx
@@ -44,7 +44,7 @@ export const useElementOptions = (filterElements?: ElementOptionsFilter | undefi
   };
 
   // default to off until we add file scanning
-  const allowFileInput = false;
+  const allowFileInput = true;
 
   const fileInputOption: ElementOption = {
     id: "fileInput",
